fixed alot of bugs
Fixed the cube.render bug which caused a compleat stop! calling and init
of the cube should now be fine. Added the right values for the
demultiplexing.
SPI fixed to const & instead of copy.
Main back to some showing like showing the temperature and so on.
(Actually the cube is already working and running here but no animation
yet!)