Major improvements
Added a Proxypattern for the cube, changed RGB to take by value instead
of reference, since the value is 8bit (uint8_t) and the reference is
16bit which makes no sense.
cube(x,y,z) = RGB is now possible.

Changed the Cubelogic to take RGB instead of single values to have a
better optimization and a more "simple" way. Moreover deleted getr getb
and so on simply take getRGB().r . No cube off while shifting new values
to the "preregister". (There is no clutter or flickting so no need to
turn them off) Increased colordepth by 1 so it can take 31 colors which
makes it %32 which is way faster then %30. (simple &31) Taking all small
values by copy instead of reference. (everything smaller then 16bit)

Fixed issue with the Display and sprintf. took a double and sprintf it
which does not work. This is fixed now.